Move NFC/memory stuff to its own header
