+++ code/headers/base_comm.hpp
@@ -6,29 +6,10 @@
 #include <type_traits>
 #include <ringbuffer.hpp>
 
+#include "priority.hpp"
 #include "frame_types.hpp"
 
 namespace r2d2 {
-    /**
-     * All priority levels that can be assigned
-     * to packets.
-     */
-    enum class priority {
-        // High priority packet
-            HIGH = 0,
-
-        // Normal priority packet, default
-            NORMAL = 1,
-
-        // Low priority packet, used when there is no hard time constraint on the delivery of the data
-            LOW = 2,
-
-        // Data stream packet. Used when a stream of packets (e.g. video data) is put on the bus.
-        // Assigning this will given these packets the lowest priority, preventing the large data stream
-        // from clogging up the bus.
-            DATA_STREAM = 3
-    };
-
     struct frame_s {
         uint8_t *data;
         size_t length;
